pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = "A simple Python-based static site generator"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.2",
    "markdown>=3.5",
    "pygments>=2.17.2",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "watchdog>=3.0.0",
]
